Modularize api
This commit is just modularization, refactoring will come later
