+++ src/pqnstack/app/api/main.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+from pqnstack.app.api.routes import chsh
+from pqnstack.app.api.routes import qkd
+from pqnstack.app.api.routes import timetagger
+
+api_router = APIRouter()
+api_router.include_router(chsh.router)
+api_router.include_router(qkd.router)
+api_router.include_router(timetagger.router)
